git-compat-util.h: fix build on gcc 4.4.x and earlier
Commit 9ff7eb8c88 uses the `deprecated` attribute to display a warning
message when something that is supposed to be unused is actually used.
However, `deprecated` only supports messages starting in gcc 4.5.0, so
earlier versions of gcc end up bailing out with a message like

  error: wrong number of arguments specified for ‘deprecated’ attribute

This commit removes the use of the `deprecated` attribute for gcc 4.4.x
and earlier. As this is just a diagnostic feature for developers, and
presumably very few people are developing on such old systems anyway,
this change shouldn't really impact anyone.

Signed-off-by: Jeremy Lin <jeremy.lin@gmail.com>